+++ DomainManagement/functions/acls/Test-DMAcl.ps1
@@ -1,5 +1,4 @@
-﻿function Test-DMAcl
-{
+﻿function Test-DMAcl {
 	<#
 		.SYNOPSIS
 			Tests whether the configured groups match a domain's configuration.
@@ -27,8 +26,7 @@
 		$Credential
 	)
 	
-	begin
-	{
+	begin {
 		$parameters = $PSBoundParameters | ConvertTo-PSFHashtable -Include Server, Credential
 		$parameters['Debug'] = $false
 		Assert-ADConnection @parameters -Cmdlet $PSCmdlet
@@ -37,6 +35,30 @@
 		Set-DMDomainContext @parameters
 
 		#region Functions
+		function New-Change {
+			[Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSUseShouldProcessForStateChangingFunctions", "")]
+			[CmdletBinding()]
+			param (
+				$Type,
+				$OldValue,
+				$NewValue,
+				[string]
+				$Identity,
+				$OldSID,
+				$NewSID
+			)
+
+			$changeItem = [PSCustomObject]@{
+				PSTypeName = 'DomainManagement.Acl.Change'
+				Type       = $Type
+				Old        = $OldValue
+				New        = $NewValue
+				Identity   = $Identity
+				OldSID     = $OldSID
+				NewSID     = $NewSID
+			}
+			Add-Member -InputObject $changeItem -MemberType ScriptMethod -Name ToString -Value { '{0}->{1}' -f $this.Type, $this.New } -Force -PassThru
+		}
 		function Get-ChangeByCategory {
 			[CmdletBinding()]
 			param (
@@ -56,32 +78,35 @@
 			$configuredSID = $Category.Owner | Resolve-String | Convert-Principal @parameters -OutputType SID
 
 			[System.Collections.ArrayList]$changes = @()
-			if ("$ownerSID" -ne "$configuredSID") { $null = $changes.Add('Owner') }
-			Compare-Property -Property NoInheritance -Configuration $Category -ADObject $aclObject -Changes $changes -ADProperty AreAccessRulesProtected
+			if ("$ownerSID" -ne "$configuredSID") {
+				$null = $changes.Add((New-Change -Identity $ADObject -Type Owner -OldValue $aclObject.Owner -NewValue ($Category.Owner | Resolve-String) -OldSID $ownerSID -NewSID $configuredSID))
+			}
+			if ($Category.NoInheritance -ne $aclObject.AreAccessRulesProtected) {
+				$null = $changes.Add((New-Change -Identity $ADObject -Type NoInheritance -OldValue $aclObject.AreAccessRulesProtected -NewValue $Category.NoInheritance))
+			}
 
 			if ($changes.Count) {
-				New-TestResult @resultDefaults -Identity $ADObject -Configuration $Category -Type Changed -Changed $changes.ToArray() -ADObject $aclObject
+				New-TestResult @resultDefaults -Identity $ADObject -Configuration $Category -Type Update -Changed $changes.ToArray() -ADObject $aclObject
 			}
 		}
 		#endregion Functions
 	}
-	process
-	{
+	process {
 		#region processing configuration
 		foreach ($aclDefinition in $script:acls.Values) {
 			$resolvedPath = Resolve-String -Text $aclDefinition.Path
 
 			$resultDefaults = @{
-				Server = $Server
-				ObjectType = 'Acl'
-				Identity = $resolvedPath
+				Server        = $Server
+				ObjectType    = 'Acl'
+				Identity      = $resolvedPath
 				Configuration = $aclDefinition
 			}
 
 			
-			if (-not (Test-ADObject @parameters -Identity $resolvedPath))  {
+			if (-not (Test-ADObject @parameters -Identity $resolvedPath)) {
 				if ($aclDefinition.Optional) { continue }
-				Write-PSFMessage -String 'Test-DMAcl.ADObjectNotFound' -StringValues $resolvedPath -Tag 'panic','failed' -Target $aclDefinition
+				Write-PSFMessage -String 'Test-DMAcl.ADObjectNotFound' -StringValues $resolvedPath -Tag 'panic', 'failed' -Target $aclDefinition
 				New-TestResult @resultDefaults -Type 'MissingADObject'
 				Continue
 			}
@@ -89,7 +114,7 @@
 			try { $aclObject = Get-AdsAcl @parameters -Path $resolvedPath -EnableException }
 			catch {
 				if ($aclDefinition.Optional) { continue }
-				Write-PSFMessage -String 'Test-DMAcl.NoAccess' -StringValues $resolvedPath -Tag 'panic','failed' -Target $aclDefinition -ErrorRecord $_
+				Write-PSFMessage -String 'Test-DMAcl.NoAccess' -StringValues $resolvedPath -Tag 'panic', 'failed' -Target $aclDefinition -ErrorRecord $_
 				New-TestResult @resultDefaults -Type 'NoAccess'
 				Continue
 			}
@@ -98,31 +123,27 @@
 			$configuredSID = $aclDefinition.Owner | Resolve-String | Convert-Principal @parameters -OutputType SID
 
 			[System.Collections.ArrayList]$changes = @()
-			if ("$ownerSID" -ne "$configuredSID") { $null = $changes.Add('Owner') }
-			Compare-Property -Property NoInheritance -Configuration $aclDefinition -ADObject $aclObject -Changes $changes -ADProperty AreAccessRulesProtected
+			if ("$ownerSID" -ne "$configuredSID") {
+				$null = $changes.Add((New-Change -Identity $resolvedPath -Type Owner -OldValue $aclObject.Owner -NewValue ($aclDefinition.Owner | Resolve-String) -OldSID $ownerSID -NewSID $configuredSID))
+			}
+			if ($aclDefinition.NoInheritance -ne $aclObject.AreAccessRulesProtected) {
+				$null = $changes.Add((New-Change -Identity $resolvedPath -Type NoInheritance -OldValue $aclObject.AreAccessRulesProtected -NewValue $aclDefinition.NoInheritance))
+			}
 
 			if ($changes.Count) {
-				New-TestResult @resultDefaults -Type Changed -Changed $changes.ToArray() -ADObject $aclObject
+				New-TestResult @resultDefaults -Type Update -Changed $changes.ToArray() -ADObject $aclObject
 			}
 		}
 		#endregion processing configuration
 
 		#region check if all ADObjects are managed
-		<#
-		Object Types ignored:
-		- Service Connection Point
-		- RID Set
-		- DFSR Settings objects
-		- Computer objects
-		Pre-defining domain controllers or other T0 servers and their meta-information objects would be an act of futility and probably harmful.
-		#>
 		$foundADObjects = foreach ($searchBase in (Resolve-ContentSearchBase @parameters -NoContainer)) {
 			Get-ADObject @parameters -LDAPFilter '(objectCategory=*)' -SearchBase $searchBase.SearchBase -SearchScope $searchBase.SearchScope
 		}
 		
 		$resolvedConfiguredPaths = $script:acls.Values.Path | Resolve-String
 		$resultDefaults = @{
-			Server = $Server
+			Server     = $Server
 			ObjectType = 'Acl'
 		}
 
@@ -137,7 +158,7 @@
 			
 			if ($script:aclByCategory.Count -gt 0) {
 				$category = Resolve-DMObjectCategory -ADObject $foundADObject @parameters
-				if ($matchingCategory = $category | Where-Object Name -in $script:aclByCategory.Keys | Select-Object -First 1) {
+				if ($matchingCategory = $category | Where-Object Name -In $script:aclByCategory.Keys | Select-Object -First 1) {
 					Get-ChangeByCategory -ADObject $foundADObject -Category $script:aclByCategory[$matchingCategory.Name] -ResultDefaults $resultDefaults -Parameters $parameters
 					continue
 				}