+++ DomainManagement/functions/groups/Invoke-DMGroup.ps1
@@ -65,7 +65,7 @@
 			}
 			
 			switch ($testItem.Type) {
-                'ShouldDelete' {
+                'Delete' {
                     Invoke-PSFProtectedCommand -ActionString 'Invoke-DMGroup.Group.Delete' -Target $testItem -ScriptBlock {
                         Remove-ADGroup @parameters -Identity $testItem.ADObject.ObjectGUID -ErrorAction Stop -Confirm:$false
                     } -EnableException $EnableException -PSCmdlet $PSCmdlet -Continue
@@ -98,9 +98,9 @@
 						}
                     } -EnableException $EnableException -PSCmdlet $PSCmdlet -Continue
                 }
-                'Changed' {
-                    if ($testItem.Changed -contains 'Path') {
-                        $targetOU = Resolve-String -Text $testItem.Configuration.Path
+                'Update' {
+                    if ($change = $testItem.Changed | Where-Object Property -eq 'Path') {
+                        $targetOU = $change.New
                         try { $null = Get-ADObject @parameters -Identity $targetOU -ErrorAction Stop }
                         catch { Stop-PSFFunction -String 'Invoke-DMGroup.Group.Update.OUExistsNot' -StringValues $testItem.Identity, $targetOU -Target $testItem -EnableException $EnableException -Continue }
 
@@ -109,15 +109,17 @@
                         } -EnableException $EnableException -PSCmdlet $PSCmdlet -Continue
                     }
                     $changes = @{ }
-                    if ($testItem.Changed -contains 'Description') { $changes['Description'] = (Resolve-String -Text $testItem.Configuration.Description) }
-                    if ($testItem.Changed -contains 'Category') { $changes['GroupCategory'] = (Resolve-String -Text $testItem.Configuration.Category) }
+					foreach ($change in $testItem.Changed | Where-Object Property -in 'Description', 'Category') {
+						if ($change.Property -eq 'Description') { $changes['Description'] = $change.New }
+						if ($change.Property -eq 'Category') { $changes['GroupCategory'] = $change.New }
+					}
 					
                     if ($changes.Keys.Count -gt 0) {
                         Invoke-PSFProtectedCommand -ActionString 'Invoke-DMGroup.Group.Update' -ActionStringValues ($changes.Keys -join ", ") -Target $testItem -ScriptBlock {
                             $null = Set-ADObject @parameters -Identity $testItem.ADObject.ObjectGUID -ErrorAction Stop -Replace $changes -Confirm:$false
                         } -EnableException $EnableException -PSCmdlet $PSCmdlet -Continue
                     }
-                    if ($testItem.Changed -contains 'Scope') {
+                    if ($testItem.Changed.Property -contains 'Scope') {
 						$targetScope = Resolve-String -Text $testItem.Configuration.Scope
 						if ($targetScope -notin ([Enum]::GetNames([Microsoft.ActiveDirectory.Management.ADGroupScope]))) {
 							Stop-PSFFunction -String 'Invoke-DMGroup.Group.InvalidScope' -StringValues $testItem, $targetScope -Continue -EnableException $EnableException -Target $testItem
@@ -128,9 +130,9 @@
 							$null = Set-ADGroup @parameters -Identity $testItem.ADObject.ObjectGUID -GroupScope $targetScope -ErrorAction Stop -Confirm:$false
                         } -EnableException $EnableException.ToBool() -PSCmdlet $PSCmdlet -Continue
 					}
-					if ($testItem.Changed -contains 'Name') {
-                        Invoke-PSFProtectedCommand -ActionString 'Invoke-DMGroup.Group.Update.Name' -ActionStringValues (Resolve-String -Text $testItem.Configuration.Name) -Target $testItem -ScriptBlock {
-                            $testItem.ADObject | Rename-ADObject @parameters -NewName (Resolve-String -Text $testItem.Configuration.Name) -ErrorAction Stop -Confirm:$false
+					if ($change = $testItem.Changed | Where-Object Property -eq 'Name') {
+                        Invoke-PSFProtectedCommand -ActionString 'Invoke-DMGroup.Group.Update.Name' -ActionStringValues $change.NewValue -Target $testItem -ScriptBlock {
+                            $testItem.ADObject | Rename-ADObject @parameters -NewName $change.New -ErrorAction Stop -Confirm:$false
                         } -EnableException $EnableException -PSCmdlet $PSCmdlet -Continue
                     }
                 }